requirements.txt
certifi==2020.12.5
jellyfish==0.8.2
mypy-extensions==0.4.3
prettytable==2.0.0
prompt-toolkit==3.0.11
Pygments==2.7.4
pyparsing==2.4.7
python-nubia==0.2b5
termcolor==1.1.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
wcwidth==0.2.5
